refactor: remove http context knowledge from business logic project
