--- Doppler.ContactPolicies.Business.Logic/UserApiClient/Handlers/TokenHandler.cs
@@ -1,26 +1,0 @@
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Http;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Doppler.ContactPolicies.Business.Logic.UserApiClient.Handlers
-{
-    public class TokenHandler : DelegatingHandler
-    {
-        private readonly IHttpContextAccessor _httpContextAccessor;
-
-        public TokenHandler(IHttpContextAccessor httpContextAccessor)
-        {
-            _httpContextAccessor = httpContextAccessor;
-        }
-        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request,
-            CancellationToken cancellationToken)
-        {
-            var token = await _httpContextAccessor.HttpContext.GetTokenAsync("Bearer", "access_token");
-            request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", token);
-            return await base.SendAsync(request, cancellationToken);
-        }
-    }
-}
