+++ Doppler.ContactPolicies.Business.Logic/UserApiClient/Services/IUsersApiTokenGetter.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace Doppler.ContactPolicies.Business.Logic.UserApiClient.Services
+{
+    public interface IUsersApiTokenGetter
+    {
+        Task<string> GetTokenAsync();
+    }
+}
